изменил структуру
